Refactor connector logic for registering proxy method and events
